% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos}
\alias{datos}
\title{conjunto de datos de ejemplo}
\format{
Los datos son generados de forma sintetica.
\describe{
\item{a}{muestra de datos}
\item{b}{muestra de datos multiplicada por cinco}
}
}
\source{
generacion propia con funciones de R
}
\usage{
datos
}
\description{
Los datos son utlilizados como ejemplo de nuestra funcion de suma.
}
\keyword{datasets}
