% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_to_C.R
\name{F_to_C}
\alias{F_to_C}
\title{Conversion de la temperatura en Fahrenheit a Centigrados}
\usage{
F_to_C(temp_fahrenheit)
}
\arguments{
\item{temp_fahrenheit}{un vector numerico de la temperatura en Fahrenheit}
}
\value{
Un vector numerico de la temperatura en grados centigrados.
}
\description{
La funcion \code{F_to_C()} toma un valor numerico de la temperatura en Fahrenheit y calcula la equivalencia en grados centigrados.
}
\details{
Estos son detalles extras de la funcion.
}
\examples{
F_to_C(100)

}
