% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros}
\usage{
suma(x, y)
}
\arguments{
\item{x}{un vector numérico}

\item{y}{un vector numérico}
}
\value{
vector númerico con la suma de \code{x} e \code{y}.
}
\description{
La funcion \code{suma()} toma 2 números positivos y los suma.
}
\details{
Estos son detalles extras de la funcion.
}
\examples{
sume(2, 2)
suma(1, 10)

}
