% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datos}
\alias{datos}
\title{Conjunto de datos de ejemplo}
\format{
Los datos son generados de manera aleatoria.
\describe{
\item{a}{muestra de datos}
\item{b}{muestra de datos multiplicada por 5}
}

los datos son generados de forma sintetica.
\describe{
\item{a}{muestra de datos}
\item{b}{muestra de datos multiplicada por cinco}
}
}
\source{
Fuente: generación propia con funciones de R.

generacion propia con funciones de R
}
\usage{
datos

datos
}
\description{
Los datos son utilizados para poder chequear nuestra función suma().

Los datos son utilzados como ejemplo de nuestra funcion de suma.
}
\keyword{datasets}
