% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suma.R
\name{suma}
\alias{suma}
\title{' La función \code{suma()} dos valores numéricos positivos y calcula la suma entre ambos números.}
\usage{
suma(x = 5, y = 2)
}
\arguments{
\item{x}{vector numerico}

\item{y}{vector numerico}
}
\value{
un numero entero
}
\description{
' La función \code{suma()} dos valores numéricos positivos y calcula la suma entre ambos números.
}
\examples{
suma(2,2)

}
