% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suma.R
\name{suma}
\alias{suma}
\title{Suma dos numeros}
\usage{
suma(x, y)
}
\arguments{
\item{x}{un vector numérico}

\item{y}{un vector numérico}
}
\value{
Un vector numerico con la suma de x e y
}
\description{
La función \code{suma()}toma dos numeros y los suma
}
\details{
Estos son detalles extras de la función
}
\examples{
suma(2,2)
suma(1,10)

}
