% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suma.R
\name{suma}
\alias{suma}
\title{La funcion \code{suma()} toma dos valores numericos positivos y calcula la suma entre ambos numeros.}
\usage{
suma(x = 5, y = 2)
}
\arguments{
\item{x}{vector numerico}

\item{y}{vector numerico}
}
\value{
un numero entero
}
\description{
La funcion \code{suma()} toma dos valores numericos positivos y calcula la suma entre ambos numeros.
}
\examples{
suma(2,2)

}
