% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suma.R
\name{suma}
\alias{suma}
\title{Title
Suma dos números}
\usage{
suma(x, y)
}
\arguments{
\item{x}{un vector numérico (debe ser positivo)}

\item{y}{un vector numérico (debe ser positivo)}
}
\value{
Un vector numérico con la suma de \code{x} e \code{y}.
}
\description{
La función \code{suma()} toma 2 números positivos y los suma
}
\details{
Estos son los detalles extras de la función.
}
\examples{
suma(2,2)
suma(1,10)


}
